Move Intent parsing to Launch Activity + only launch playback activity using the launch activity
