+++ app/src/main/AndroidManifest.xml
@@ -2,8 +2,10 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="de.shadow578.yetanothervideoplayer">
     <!-- To adjust Media Volume by swiping -->
-    <uses-permission android:name="android.permission.MODIFY_AUDIO_SETTINGS" /> <!-- To Stream Video -->
-    <uses-permission android:name="android.permission.INTERNET" /> <!-- To Play local files -->
+    <uses-permission android:name="android.permission.MODIFY_AUDIO_SETTINGS" />
+    <!-- To Stream Video -->
+    <uses-permission android:name="android.permission.INTERNET" />
+    <!-- To Play local files -->
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 
     <application
@@ -14,11 +16,14 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:usesCleartextTraffic="true">
-        <activity android:name=".ui.LaunchActivity"></activity>
         <!-- ^^allow traffic over http, since not all streams will be via https -->
+
+        <!-- Settings Activity -->
         <activity
             android:name=".ui.AppSettingsActivity"
             android:label="@string/settings_root_title" />
+
+        <!-- Video Test Activity for Debugging -->
         <activity
             android:name=".ui.VideoTestActivity"
             android:label="@string/app_name">
@@ -28,13 +33,21 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <!-- Playback Activity (only started internally using LaunchActivity -->
         <activity
             android:name=".ui.PlaybackActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenSize|screenLayout|smallestScreenSize|uiMode"
             android:label="@string/app_name"
             android:resizeableActivity="true"
             android:supportsPictureInPicture="true"
-            android:theme="@style/AppTheme.PlaybackActivity">
+            android:theme="@style/AppTheme.PlaybackActivity"/>
+
+        <!-- Launch Activity (Splash Screen -->
+        <activity
+            android:name=".ui.LaunchActivity"
+            android:label="@string/app_name"
+            android:theme="@style/AppTheme.SplashActivity">
 
             <!-- share with yavp -->
             <intent-filter android:label="@string/app_open_with">
@@ -217,5 +230,4 @@
             </intent-filter>
         </activity>
     </application>
-
 </manifest>
